Added file in separate topic folder

+++ 2-variables/Int data type/int-variable.c
@@ -1,6 +1,7 @@
 #include<stdio.h>
 int main(){
 	// Integer Variable
-	int num = 11;
+	unsigned long int num = 11;
 	printf("%i",num);
 }
+
